requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
openai>=1.58.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
pydantic>=2.0.0,<3.0.0
httpx>=0.27.0,<1.0.0
pytest>=8.0.0,<9.0.0
pytest-asyncio>=0.24.0,<1.0.0
tenacity>=8.0.0,<9.0.0
opentelemetry-api>=1.20.0,<2.0.0
opentelemetry-sdk>=1.20.0,<2.0.0
opentelemetry-exporter-otlp>=1.20.0,<2.0.0
opentelemetry-instrumentation-openai>=0.1.0,<1.0.0
numpy>=1.24.0,<2.0.0
tiktoken>=0.7.0,<1.0.0
